now It's working version

+++ src/main/java/com/abdala/demo/controller/UserController.java
@@ -44,20 +44,20 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserDTO> getUserById(@PathVariable int id) {
+    public ResponseEntity<UserDTO> getUserById(@PathVariable Long id) {
         UserDTO userDTO = userService.getUserById(id);
         return ResponseEntity.ok(userDTO);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<UserDTO>  updateUser(@PathVariable int  id, @RequestBody UpdateUserDTO updateUserDTO) {
+    public ResponseEntity<UserDTO>  updateUser(@PathVariable Long  id, @RequestBody UpdateUserDTO updateUserDTO) {
 
         UserDTO updatedUser = userService.updateUser(id, updateUserDTO);
         return new ResponseEntity<>(updatedUser, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable int id) {
+    public void deleteUser(@PathVariable Long id) {
         userService.deleteUser(id);
     }
